+++ .editorconfig
@@ -8,7 +8,7 @@ end_of_line = lf
 trim_trailing_whitespace = true
 insert_final_newline = true
 
-[{.,}*.{ts{,*},js{,*},y{a,}ml}]
+[*.{ts,tsx,js,yaml,yml}]
 indent_style = space
 indent_size = 2
