+++ package.json
@@ -55,11 +55,11 @@
     "_postinstall": "husky install",
     "clean": "rimraf dist coverage",
     "lint:fix": "npm run -s lint -- --fix",
-    "lint": "eslint source/**/*.js",
+    "lint": "eslint src/**/*.ts",
     "postpublish": "pinst --enable",
     "prepublishOnly": "pinst --disable",
     "semantic-release": "semantic-release",
-    "start": "ts-node src/cli/index.ts",
+    "start": "node -r ts-node/register src/cli/index.ts",
     "tdd": "npm run -s test -- --watch",
     "test:ci": "npm run -s test",
     "test": "NODE_ENV=test jest"
