+++ src/cli/index-stock.ts
@@ -1,0 +1,7 @@
+import { program } from './lib/program';
+
+program
+  .action(() => {
+    console.log('stock');
+  })
+  .parseAsync();
